test(RM): parallel testing
Splits the testing of single components of realm managements into
separate modules to allow for parallel testing
